chg: Test of VisualStudio 2005 Beta2 for compability
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%40368

+++ Altaxo/Base/Graph/Axes/Log10Axis.cs
@@ -326,7 +326,7 @@ namespace Altaxo.Graph.Axes
       // calculate the number of major ticks
 
       int nFullDecades = (int)(1+Math.Floor(log10end)-Math.Ceiling(log10org));
-      int nMajorTicks = (int)Math.Floor((nFullDecades+m_DecadesPerMajorTick-1)/m_DecadesPerMajorTick);
+      int nMajorTicks = (int)Math.Floor((nFullDecades+m_DecadesPerMajorTick-1)/(double)m_DecadesPerMajorTick);
 
 
       double[] retval = new double[nMajorTicks];
